refactor to match context best practices
