+++ scm-ui/ui-webapp/src/containers/Index.tsx
@@ -21,9 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import React, { FC, useCallback, useState } from "react";
+import React, { FC, useState } from "react";
 import App from "./App";
-import { ActiveModalCountContext, ErrorBoundary, Header, Loading } from "@scm-manager/ui-components";
+import { ActiveModalCountContextProvider, ErrorBoundary, Header, Loading } from "@scm-manager/ui-components";
 import PluginLoader from "./PluginLoader";
 import ScrollToTop from "./ScrollToTop";
 import IndexErrorPage from "./IndexErrorPage";
@@ -37,9 +37,6 @@ import InitializationPluginWizardStep from "./InitializationPluginWizardStep";
 const Index: FC = () => {
   const { isLoading, error, data } = useIndex();
   const [pluginsLoaded, setPluginsLoaded] = useState(false);
-  const [activeModalCount, setActiveModalCount] = useState(0);
-  const incrementModalCount = useCallback(() => setActiveModalCount((prev) => prev + 1), []);
-  const decrementModalCount = useCallback(() => setActiveModalCount((prev) => prev - 1), []);
 
   // TODO check componentDidUpdate method for anonymous user stuff
 
@@ -61,19 +58,17 @@ const Index: FC = () => {
 
   const link = (data._links.uiPlugins as Link).href;
   return (
-    <ActiveModalCountContext.Provider
-      value={{ value: activeModalCount, increment: incrementModalCount, decrement: decrementModalCount }}
-    >
-      <ErrorBoundary fallback={IndexErrorPage}>
-        <ScrollToTop>
+    <ErrorBoundary fallback={IndexErrorPage}>
+      <ScrollToTop>
+        <ActiveModalCountContextProvider>
           <NamespaceAndNameContextProvider>
             <PluginLoader link={link} loaded={pluginsLoaded} callback={() => setPluginsLoaded(true)}>
               <App />
             </PluginLoader>
           </NamespaceAndNameContextProvider>
-        </ScrollToTop>
-      </ErrorBoundary>
-    </ActiveModalCountContext.Provider>
+        </ActiveModalCountContextProvider>
+      </ScrollToTop>
+    </ErrorBoundary>
   );
 };
